requirements.txt
absl-py==2.0.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
argcomplete==3.0.8
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.7.0
blosc2==2.4.0
cached-property
cachetools==5.3.2
category-encoders==2.6.3
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
cloudpickle==3.0.0
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cssutils==2.9.0
cxxfilt==0.3.0
cycler==0.11.0
Cython==3.0.8
dacite==1.8.1
dataframe-image==0.2.3
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
entrypoints 
exceptiongroup 
executing==1.2.0
fastjsonschema==2.17.1
flake8==6.0.0
flatbuffers==23.5.26
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.4.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
gobject==0.1.0
google-auth==2.25.1
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
html2image==2.0.4.3
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
imageio==2.33.1
imbalanced-learn==0.11.0
importlib-metadata==6.8.0
importlib-resources 
iniconfig==2.0.0
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.1.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.23.0
jupyterthemes==0.20.0
keras==2.15.0
kiwisolver==1.4.4
lazy_loader==0.3
lesscpy==0.15.1
libclang==16.0.6
llvmlite==0.41.1
lxml==5.1.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
missingno==0.5.2
mistune==3.0.1
ml-dtypes==0.2.0
mrmr-selection==0.2.8
msgpack==1.0.7
multidict==6.0.5
multimethod==1.10
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.1
ndindex==1.7
nest-asyncio==1.5.6
networkx==3.2.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.58.1
numexpr==2.8.8
numpy==1.25.1
oauthlib==3.2.2
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.1
pandas==2.1.4
pandas-profiling==3.2.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.5
phik==0.12.3
pickleshare==0.7.5
Pillow==10.0.0
pipx==1.2.0
pkgutil_resolve_name
platformdirs==3.8.1
pluggy==1.3.0
ply==3.11
polars==0.20.6
prometheus-client==0.17.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.23.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
pyflakes==3.0.1
Pygments==2.15.1
pykwalify==1.8.0
pympress==1.8.2
pymrmr==0.1.11
pyparsing==3.0.9
pyradiomics==3.0.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz
PyWavelets==1.5.0
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzbar==0.1.9
pyzmq==25.1.0
referencing==0.29.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.8.10
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.12.2
Send2Trash==1.8.2
shap==0.44.0
SimpleITK==2.3.1
six==1.16.0
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
statsmodels==0.14.1
streamlit==1.28.2
tables==3.9.2
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.12.9
tinycss2==1.2.1
tk==0.1.0
toml==0.10.2
tomli
toolz==0.12.0
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
typeguard==4.1.5
types-python-dateutil
typing-utils
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
urllib3==2.0.3
userpath==1.8.0
validators==0.22.0
visions==0.7.5
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==3.0.1
wordcloud==1.9.3
wrapt==1.14.1
yarl==1.9.4
ydata-profiling==4.6.3
zipp
